+++ cpp/devices/disk_cache.h
@@ -27,7 +27,7 @@ using namespace piscsi_interface;
 class DiskCache
 {
 	// Number of tracks to cache
-	static const int CACHE_MAX = 16;
+	static const int64_t CACHE_MAX = 16;
 
 	uint64_t read_error_count = 0;
 	uint64_t write_error_count = 0;
@@ -47,23 +47,23 @@ public:
 		uint32_t serial;				// Serial
 	};
 
-	DiskCache(const string&, int, uint32_t, off_t = 0);
+	DiskCache(const string&, int, uint64_t, off_t = 0);
 	~DiskCache() = default;
 
 	void SetRawMode(bool b) { cd_raw = b; }		// CD-ROM raw mode setting
 
 	bool Save();							// Save and release all
-	bool ReadSector(span<uint8_t>, uint32_t);			// Sector Read
-	bool WriteSector(span<const uint8_t>, uint32_t);	// Sector Write
+	bool ReadSector(span<uint8_t>, uint64_t);			// Sector Read
+	bool WriteSector(span<const uint8_t>, uint64_t);	// Sector Write
 
 	vector<PbStatistics> GetStatistics(bool) const;
 
 private:
 
 	// Internal Management
-	shared_ptr<DiskTrack> Assign(int);
-	shared_ptr<DiskTrack> GetTrack(uint32_t);
-	bool Load(int index, int track, shared_ptr<DiskTrack>);
+	shared_ptr<DiskTrack> Assign(int64_t);
+	shared_ptr<DiskTrack> GetTrack(uint64_t);
+	bool Load(int index, int64_t track, shared_ptr<DiskTrack>);
 	void UpdateSerialNumber();
 
 	// Internal data
@@ -71,7 +71,7 @@ private:
 	uint32_t serial = 0;						// Last serial number
 	string sec_path;							// Path
 	int sec_size;								// Sector Size (8=256, 9=512, 10=1024, 11=2048, 12=4096)
-	int sec_blocks;								// Blocks per sector
+	int64_t sec_blocks;								// Blocks per sector
 	bool cd_raw = false;						// CD-ROM RAW mode
 	off_t imgoffset;							// Offset to actual data
 };
